requirements.txt
1、Django下载地址：
https://www.djangoproject.com/download/
版本选择：1.11.14
# pip install Django==1.11.14


2、django-rest-framework下载地址：
http://www.django-rest-framework.org/
版本选择：3.8.2
# pip install djangorestframework==3.8.2


3、django-session-security下载地址：
https://pypi.org/project/django-session-security/
版本选择：2.6.0
# pip install django-session-security==2.6.0


4、cryptography下载地址：
https://pypi.org/project/cryptography/
版本选择：2.3
# pip install cryptography==2.3


5、paramiko下载地址：
http://www.paramiko.org/installing.html
版本选择：2.0.2
# pip install paramiko==2.0.2


6、pycparser下载地址：
https://pypi.org/project/pycparser/#files
版本选择：2.18
# pip install pycparser==2.18


7、PyMySQL下载地址：
https://pypi.org/project/PyMySQL/#description
版本选择：0.9.2
# pip install PyMySQL==0.9.2


8、sh
sh==1.12.9
# pip install sh==1.12.9


9、mysqlclient
要求先：# yum install python-devel mysql-devel
mysqlclient==1.3.12
# pip install mysqlclient==1.3.12


10、pytest
pytest==3.0.7
# pip install pytest==3.0.7


11、celery
celery==4.2.0
# pip install celery==4.2.0


12、django-celery-beat
django-celery-beat==1.1.1
# pip install django-celery-beat==1.1.1


13、gunicorn
gunicorn==19.7.1
# pip install gunicorn==19.7.1


14、redis
redis==2.10.5
# pip install redis==2.10.5


15、psutil
psutil==5.2.2
# pip install psutil==5.2.2


16、schedule
schedule==0.4.3
# pip install schedule==0.4.3


17、pymongo
pymongo==3.3.0
# pip install pymongo==3.3.0


18、requests
requests==2.11.1
# pip install requests==2.11.1


19、django-db
django-db==0.0.7
# pip install django-db==0.0.7


20、pillow
# pip install Pillow==5.2.0
